add session token authentication to all neccessary handlers

+++ src/handlers.rs
@@ -1,31 +1,11 @@
-use actix_web::{web, HttpResponse, Responder};
+use actix_web::{web, HttpResponse, HttpRequest, Responder};
 use std::sync::{Arc, Mutex};
 use crate::db::CouchDB;
 use serde_json::Value;
 use crate::auth::UserManager;
 use crate::email::EmailManager;
 use serde::Deserialize;
-
-pub async fn get_document(db: web::Data<Arc<CouchDB>>, id: web::Path<String>) -> impl Responder {
-    println!("Get_document handler");
-    match db.get_document_data(&id).await {
-        Ok(doc) => HttpResponse::Ok().json(doc),
-        Err(e) => {
-            println!("Error: {:?}", e);
-            HttpResponse::NotFound().body(format!("Document with id {} not found", id))
-        }
-    }
-}
-
-pub async fn put_document(db: web::Data<Arc<CouchDB>>, id: web::Path<String>, data: web::Json<Value>) -> impl Responder {
-    match db.put_document(&id, data.into_inner()).await {
-        Ok(doc) => HttpResponse::Ok().json(doc),
-        Err(e) => {
-            println!("Error: {:?}", e);
-            HttpResponse::InternalServerError().body("Internal Server Error")
-        }
-    }
-}
+use crate::utils;
 
 #[derive(Deserialize)]
 pub struct LoginData {
@@ -61,6 +41,61 @@ pub struct AddUuid {
     uuid: String
 }
 
+
+
+pub async fn get_document(user_manager: web::Data<Arc<Mutex<UserManager>>>, db: web::Data<Arc<CouchDB>>, id: web::Path<String>, req: HttpRequest) -> impl Responder {
+    // Check if valid session token
+    let email = match utils::get_request_context(req, user_manager.clone()) {
+        Ok(context) => context.email,
+        Err(e) => return e,
+    };
+    // Check if correct session token
+    let user_manager = match user_manager.lock() {
+        Ok(manager) => manager,
+        Err(_) => return HttpResponse::InternalServerError().body("Internal Server Error"),
+    };
+    let uuids = match user_manager.get_user(&email) {
+        Some(user) => &user.uuids,
+        None => return HttpResponse::NotFound().body("User not found"),
+    };
+    if !uuids.contains(&id.to_string()) {
+        return HttpResponse::Unauthorized().body("Unauthorized");
+    }
+    match db.get_document_data(&id).await {
+        Ok(doc) => HttpResponse::Ok().json(doc),
+        Err(e) => {
+            println!("Error: {:?}", e);
+            HttpResponse::NotFound().body(format!("Document with id {} not found", id))
+        }
+    }
+}
+
+pub async fn put_document(user_manager: web::Data<Arc<Mutex<UserManager>>>, db: web::Data<Arc<CouchDB>>, id: web::Path<String>, data: web::Json<Value>, req: HttpRequest) -> impl Responder {
+    // Check if valid session token
+    let email = match utils::get_request_context(req, user_manager.clone()) {
+        Ok(context) => context.email,
+        Err(e) => return e,
+    };
+    // Check if correct session token
+    let user_manager = match user_manager.lock() {
+        Ok(manager) => manager,
+        Err(_) => return HttpResponse::InternalServerError().body("Internal Server Error"),
+    };
+    let uuids = match user_manager.get_user(&email) {
+        Some(user) => &user.uuids,
+        None => return HttpResponse::NotFound().body("User not found"),
+    };
+    if !uuids.contains(&id.to_string()) {
+        return HttpResponse::Unauthorized().body("Unauthorized");
+    }
+    match db.put_document(&id, data.into_inner()).await {
+        Ok(doc) => HttpResponse::Ok().json(doc),
+        Err(e) => {
+            println!("Error: {:?}", e);
+            HttpResponse::InternalServerError().body("Internal Server Error")
+        }
+    }
+}
 
 pub async fn login(db: web::Data<Arc<CouchDB>>, auth_data: web::Json<LoginData>, user_manager: web::Data<Arc<Mutex<UserManager>>>) -> impl Responder {
     let mut user_manager = match user_manager.lock() {
@@ -136,7 +171,16 @@ pub async fn pre_register(auth_data: web::Json<PreRegisterData>, db: web::Data<A
     }
 }
 
-pub async fn get_uuids(db: web::Data<Arc<CouchDB>>, id: web::Path<String>) -> impl Responder {
+pub async fn get_uuids(user_manager: web::Data<Arc<Mutex<UserManager>>>, db: web::Data<Arc<CouchDB>>, id: web::Path<String>, req: HttpRequest) -> impl Responder {
+    // Check if valid session token
+    let email = match utils::get_request_context(req, user_manager.clone()) {
+        Ok(context) => context.email,
+        Err(e) => return e,
+    };
+    //Check if correct session token
+    if !email.eq(&id.to_string()) {
+        return HttpResponse::Unauthorized().body("Unauthorized");
+    }
     match db.get_user(&id).await {
         Ok(user) => HttpResponse::Ok().json(user.uuids),
         Err(e) => {
@@ -146,7 +190,16 @@ pub async fn get_uuids(db: web::Data<Arc<CouchDB>>, id: web::Path<String>) -> im
     }
 }
 
-pub async fn post_uuid(db: web::Data<Arc<CouchDB>>, id: web::Path<String>, data: web::Json<AddUuid>) -> impl Responder {
+pub async fn post_uuid(user_manager: web::Data<Arc<Mutex<UserManager>>>, req: HttpRequest, db: web::Data<Arc<CouchDB>>, id: web::Path<String>, data: web::Json<AddUuid>) -> impl Responder {
+    // Check if valid session token
+    let email = match utils::get_request_context(req, user_manager.clone()) {
+        Ok(context) => context.email,
+        Err(e) => return e,
+    };
+    //Check if correct session token
+    if !email.eq(&id.to_string()) {
+        return HttpResponse::Unauthorized().body("Unauthorized");
+    }
     let mut user = match db.get_user(&id).await {
         Ok(user) => user,
         Err(e) => {
@@ -164,8 +217,17 @@ pub async fn post_uuid(db: web::Data<Arc<CouchDB>>, id: web::Path<String>, data:
     }
 }
 
-pub async fn delete_uuid(db: web::Data<Arc<CouchDB>>, path: web::Path<(String, String)>) -> impl Responder {
+pub async fn delete_uuid(user_manager: web::Data<Arc<Mutex<UserManager>>>, req: HttpRequest, db: web::Data<Arc<CouchDB>>, path: web::Path<(String, String)>) -> impl Responder {
     let (id, uuid) = path.into_inner();
+    // Check if valid session token
+    let email = match utils::get_request_context(req, user_manager.clone()) {
+        Ok(context) => context.email,
+        Err(e) => return e,
+    };
+    //Check if correct session token
+    if !email.eq(&id.to_string()) {
+        return HttpResponse::Unauthorized().body("Unauthorized");
+    }
     let mut user = match db.get_user(&id).await {
         Ok(user) => user,
         Err(e) => {
@@ -240,7 +302,16 @@ pub async fn reset_password(data: web::Json<ResetData>, user_manager: web::Data<
     HttpResponse::NotFound().body("No email found for this uuid")
 }
 
-pub async fn delete_user(email: web::Path<String> , user_manager: web::Data<Arc<Mutex<UserManager>>>, db: web::Data<Arc<CouchDB>>) -> impl Responder {
+pub async fn delete_user(req: HttpRequest, email: web::Path<String> , user_manager: web::Data<Arc<Mutex<UserManager>>>, db: web::Data<Arc<CouchDB>>) -> impl Responder {
+    // Check if valid session token
+    let _email = match utils::get_request_context(req, user_manager.clone()) {
+        Ok(context) => context.email,
+        Err(e) => return e,
+    };
+    //Check if correct session token
+    if !_email.eq(&email.to_string()) {
+        return HttpResponse::Unauthorized().body("Unauthorized");
+    }
     let user_db_deleted = match db.delete_user(&email.as_str()).await {
         Ok(_) => true,
         Err(_) => false
